+++ packages/hyde/README.md
