+++ packages/llm/providers/google/google.go
@@ -2,12 +2,12 @@ package google
 
 import (
 	"context"
+	"github.com/stillmatic/gollum/packages/llm"
 	"log/slog"
 	"strings"
 
 	"github.com/google/generative-ai-go/genai"
 	"github.com/pkg/errors"
-	"github.com/stillmatic/gollum/llm"
 	"google.golang.org/api/iterator"
 )
